+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Added
 - Rule enforcing that passwords were not set during a specific period.
 - German translations.
+- Possible to specify minimum weight to consider when testing policy.
 
 ### Changed
 - By default, test all rules regardless of weight.